Use Logger instead of puts for logs

+++ lib/tagrity/commands/start.rb
@@ -2,6 +2,7 @@ require 'listen'
 require 'tagrity/pid_file'
 require 'tagrity/helper'
 require 'tagrity/provider'
+require 'tagrity/tlogger'
 
 module Tagrity
   module Command
@@ -18,24 +19,34 @@ module Tagrity
           tag_generator = Provider.provide(:tag_generator)
           PidFile.write(PidFile.new(dir, Process.pid))
 
+          logger.fg = fg
+          logger.info("Watching #{dir} with process pid #{Process.pid}")
+
           tag_generator.generate_all if fresh
 
           listener = Listen.to(
             dir,
             relative: true,
           ) do |modified, added, removed|
-            puts "modified absolute path: #{modified}"
-            puts "added absolute path: #{added}"
-            puts "removed absolute path: #{removed}"
-            tag_generator.generate(modified)
-            tag_generator.generate(added)
-            tag_generator.delete_files_tags(removed)
+            unless modified.empty?
+              logger.info("modified absolute path: #{modified}")
+              tag_generator.generate(modified)
+            end
+            unless added.empty?
+              logger.info("added absolute path: #{added}")
+              tag_generator.generate(added)
+            end
+            unless removed.empty?
+              logger.info("removed absolute path: #{removed}")
+              tag_generator.delete_files_tags(removed)
+            end
           end
           listener.start
           sleep
         rescue ErrorProcessAlreadyRunning => e
-          puts e.message
+          logger.error(e.message)
         rescue Interrupt => e
+          logger.info("Process interrupted. Killing #{Process.pid}")
           PidFile.delete(dir)
         end
 
@@ -47,6 +58,10 @@ module Tagrity
             pids = running_processes.map { |pid_file| pid_file.pid }
             raise ErrorProcessAlreadyRunning, "Error: tagrity is already watching #{dir} with process #{pids}"
           end
+        end
+
+        def logger
+          @logger ||= Tlogger.instance
         end
       end
     end
